bot/requirements.txt
highrise-bot-sdk==24.1.0
requests==2.31.0
spotipy==2.23.0
youtube-dl==2021.12.17
soundcloud-python==0.5.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
asyncio
python-dotenv==1.0.0
psycopg2-binary==2.9.7
